separation of concerns - extract containers and components
