+++ src/routes.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Route, Redirect } from 'react-router-dom';
-import Header from './views/header/index';
-import Servers from './views/servers/index';
-import Login from './views/login/index';
+import Header from './views/header/container';
+import Servers from './views/servers/container';
+import Login from './views/login/container';
 
 const PrivateRoutes = connect(state => (
   { token: state.ui.token }
